Modified classes to google's style standards Documented ZSKeyValuePair
In our own codebase, we do not use @private, and rely on our programmers to use accessors across the board.

The reason being, if you want to extend a class, you now have to modify the parent class, which is a big no-no.  I'll do it in this library since it has the advantage of preventing invalid usage if programmers aren't aware of our convention.